# docs/swagger.yaml
definitions:
...
  main.Seat:
    properties:
      hall_id:
        type: string
      id:
        type: string
      row_number:
        type: integer
      seat_number:
        type: integer
      seat_type_id:
        type: string
    type: object
  main.SeatType:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  main.Ticket:
    properties:
      id:
        type: string
      movie_show_id:
        type: string
      price:
        type: number
      seat_id:
        type: string
      ticket_status_id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample
  title: Your API Title
  version: "1.0"
paths:
...
  /seat-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.SeatType'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все типы мест
      tags:
      - seat-types
    post:
      consumes:
      - application/json
      parameters:
      - description: Тип места
        in: body
        name: seat_type
        required: true
        schema:
          $ref: '#/definitions/main.SeatType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.SeatType'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать тип места
      tags:
      - seat-types
  /seat-types/{id}:
    delete:
      parameters:
      - description: ID типа места
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Тип не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить тип места
      tags:
      - seat-types
    get:
      parameters:
      - description: ID типа места
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SeatType'
        "404":
          description: Тип места не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить тип места по ID
      tags:
      - seat-types
    put:
      consumes:
      - application/json
      parameters:
      - description: ID типа места
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные типа
        in: body
        name: seat_type
        required: true
        schema:
          $ref: '#/definitions/main.SeatType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SeatType'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Тип не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить тип места
      tags:
      - seat-types
...
  /seats/{id}:
    delete:
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Место не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить место
      tags:
      - seats
    get:
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Seat'
        "404":
          description: Место не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить место по ID
      tags:
      - seats
    put:
      consumes:
      - application/json
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные места
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/main.Seat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Seat'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Место не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить место
      tags:
      - seats
  /seats/{movie_show_id}:
    get:
      parameters:
      - description: ID показа фильма
        in: path
        name: movie_show_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Ticket'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все билеты для показа фильма по ID
      tags:
      - seats
...
  /tickets:
    post:
      consumes:
      - application/json
      parameters:
      - description: Билет
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/main.Ticket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Ticket'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать билет
      tags:
      - tickets
  /tickets/{id}:
    delete:
      parameters:
      - description: ID билета
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить билет
      tags:
      - tickets
    get:
      parameters:
      - description: ID билета
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Ticket'
        "404":
          description: Билет не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить билет по ID
      tags:
      - tickets
    put:
      consumes:
      - application/json
      parameters:
      - description: ID билета
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённый билет
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/main.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Ticket'
        "404":
          description: Билет не найден
          schema:
            type: string
        "500":
          description: Ошибка
          schema:
            type: string
      summary: Обновить билет
      tags:
      - tickets
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"